requirements.txt
# -------- Core --------
fastapi==0.110.0
uvicorn==0.27.1

# -------- Config / Env --------
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# -------- Multipart / Upload --------
python-multipart==0.0.9

# -------- Security / Auth --------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# -------- Database --------
pymongo==4.6.1

# -------- Computer Vision --------
numpy==1.26.4
opencv-python==4.9.0.80

# -------- Face Recognition --------
insightface==0.7.3
onnxruntime==1.17.1
